Move BlogSlide component to folder
