+++ hs-web-app/src/Blogs/Component/BlogSlide/BlogSlide.js
